+++ test/samTextViewer/GenomicCoordsHistoryTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -15,6 +16,71 @@ import exceptions.InvalidGenomicCoordsException;
 
 public class GenomicCoordsHistoryTest {
 
+	@Test
+	public void canReadHistoricPositions() throws InvalidGenomicCoordsException, IOException{
+
+		// This is the initial coordinates. Set by the user or reading the input files.
+		// The historic positions will be checked against this coordinates.
+		GenomicCoords checkGc= new GenomicCoords("chr1:1-10000", null, "test_data/chr7.fa");
+
+		GenomicCoordsHistory gch= new GenomicCoordsHistory();
+		gch.readHistory(new File("test_data/asciigenome_history"), checkGc);
+		assertEquals(2, gch.getHistory().size());
+		assertEquals("chr7:10-1000", gch.getHistory().get(0).toStringRegion());
+		
+		// Initial genomic coords has no sequence dict, so everything in the history
+		// file is loaded:
+		checkGc= new GenomicCoords("chr1:1-10000", null, null);
+
+		gch= new GenomicCoordsHistory();
+		gch.readHistory(new File("test_data/asciigenome_history"), checkGc);
+		assertEquals(4, gch.getHistory().size());
+		assertEquals("bar:10-1000", gch.getHistory().get(0).toStringRegion());
+
+		// Behavior with missing or file
+		gch= new GenomicCoordsHistory();
+		gch.readHistory(new File("test_data/nonsense"), checkGc);
+	}
+	
+	@Test
+	public void canGetHistoryAsStringForHistoryFile() throws InvalidGenomicCoordsException, IOException{
+
+		GenomicCoordsHistory gch= new GenomicCoordsHistory();
+		// History is mixture of positions read from file, some of which invalid, and
+		// positions visited in this session.
+		gch.readHistory(new File("test_data/asciigenome_history"), new GenomicCoords("chr7:1-100", null, "test_data/chr7.fa"));
+		GenomicCoords g1= new GenomicCoords("chr7:1-100", null, "test_data/chr7.fa");
+		GenomicCoords g2= new GenomicCoords("chr7:2-100", null, "test_data/chr7.fa");
+		GenomicCoords g3= new GenomicCoords("chr7:3-100", null, "test_data/chr7.fa");
+		GenomicCoords g4= new GenomicCoords("chr7:4-100", null, "test_data/chr7.fa");
+		GenomicCoords g5= new GenomicCoords("chr7:5-100", null, "test_data/chr7.fa");
+		gch.add(g1);
+		gch.add(g2);
+		gch.add(g3);
+		gch.add(g4);
+		gch.add(g5);
+		
+		int maxPos= 100;
+		List<String> hist= gch.prepareHistoryForHistoryFile(new File("test_data/asciigenome_history"), maxPos);
+		assertEquals(9, hist.size()); // 2 valid pos from history + 2 invalid + 5 from current session
+		assertEquals("## pos ##bar:10-1000", hist.get(0));
+		assertEquals("## pos ##chr7:5-100", hist.get(hist.size()-1));
+
+		maxPos= 2;
+		hist= gch.prepareHistoryForHistoryFile(new File("test_data/asciigenome_history"), maxPos);
+		assertEquals(maxPos, hist.size());
+		assertEquals("## pos ##chr7:4-100", hist.get(0));
+		assertEquals("## pos ##chr7:5-100", hist.get(1));
+
+		maxPos= 0; // Don't write history
+		hist= gch.prepareHistoryForHistoryFile(new File("test_data/asciigenome_history"), maxPos);
+		assertEquals(0, hist.size());
+		
+		// Behave nicely with missing file
+		hist= gch.prepareHistoryForHistoryFile(new File("test_data/nonsense"), 100);
+
+	}
+	
 	@Test
 	public void canMoveBackAndForthInHistory() throws InvalidGenomicCoordsException, IOException{